tests/requirements.txt
pytest>=6.2, <7
python-slugify>=5.0, <6
